+++ examples/data-sources/iosxe_spanning_tree/data-source.tf
@@ -1,0 +1,2 @@
+data "iosxe_spanning_tree" "example" {
+}