+++ examples/resources/iosxe_spanning_tree/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_spanning_tree.example "Cisco-IOS-XE-native:native/spanning-tree"